feat: HostInfoCardWithLink 컴포넌트에서 NavigateBox 분리
